Add classes to own files
