+++ src/app.js
@@ -1,311 +1,11 @@
 import '@fortawesome/fontawesome-free/js/all';
 import { Modal, Collapse } from 'bootstrap';
+import CalorieTracker from './js/Tracker';
+import Meal from './js/Meal';
+import Workout from './js/Workout';
+
 import './assets/css/bootstrap.css';
 import './css/style.css';
-
-class CalorieTracker {
-  constructor() {
-    this._calorieLimit = Storage.getCalorieLimit();
-    this._totalCalories = Storage.getTotalCalories();
-    this._meals = Storage.getMeals();
-    this._workouts = Storage.getWorkouts();
-
-    this._displayCaloriesLimit();
-    this._displayCaloriesTotal();
-    this._displayCaloriesConsumed();
-    this._displayCaloriesBurned();
-    this._displayCaloriesRemaining();
-    this._displayCaloriesProgress();
-
-    document.getElementById('limit').value = this._calorieLimit;
-  }
-
-  // Public Methods
-  addMeal(meal) {
-    this._meals.push(meal);
-    this._totalCalories += meal.calories;
-    Storage.updateTotalCalories(this._totalCalories);
-    Storage.saveMeal(meal);
-    this._displayNewMeal(meal);
-    this._render();
-  }
-
-  removeMeal(id) {
-    const index = this._meals.findIndex((meal) => meal.id === id);
-    if (index !== -1) {
-      const meal = this._meals[index];
-      this._totalCalories -= meal.calories;
-      Storage.updateTotalCalories(this._totalCalories);
-      this._meals.splice(index, 1);
-      Storage.removeMeal(id);
-      this._render();
-    }
-  }
-
-  addWorkout(workout) {
-    this._workouts.push(workout);
-    this._totalCalories -= workout.calories;
-    Storage.updateTotalCalories(this._totalCalories);
-    Storage.saveWorkout(workout);
-    this._displayNewWorkout(workout);
-    this._render();
-  }
-
-  removeWorkout(id) {
-    const index = this._workouts.findIndex((workout) => workout.id === id);
-    if (index !== -1) {
-      const workout = this._workouts[index];
-      this._totalCalories += workout.calories;
-      Storage.updateTotalCalories(this._totalCalories);
-      this._workouts.splice(index, 1);
-      Storage.removeWorkout(id);
-      this._render();
-    }
-  }
-
-  reset() {
-    this._totalCalories = 0;
-    this._meals = [];
-    this._workouts = [];
-    Storage.clearAll();
-    this._render();
-  }
-
-  setLimit(calorieLimit) {
-    this._calorieLimit = calorieLimit;
-    Storage.setCalorieLimit(calorieLimit);
-    this._displayCaloriesLimit();
-    this._render();
-  }
-
-  loadItems() {
-    this._meals.forEach((meal) => this._displayNewMeal(meal));
-    this._workouts.forEach((workout) => this._displayNewWorkout(workout));
-  }
-
-  //   Private Methods
-  _displayCaloriesTotal() {
-    const totalCaloriesEl = document.getElementById('calories-total');
-    totalCaloriesEl.innerHTML = this._totalCalories;
-  }
-
-  _displayCaloriesLimit() {
-    const calorieLimitEl = document.getElementById('calories-limit');
-    calorieLimitEl.innerHTML = this._calorieLimit;
-  }
-
-  _displayCaloriesConsumed() {
-    const caloriesConsumedEl = document.getElementById('calories-consumed');
-
-    caloriesConsumedEl.innerHTML = this._meals.reduce((total, meal) => {
-      return total + meal.calories;
-    }, 0);
-  }
-
-  _displayCaloriesBurned() {
-    const caloriesBurnedEl = document.getElementById('calories-burned');
-
-    caloriesBurnedEl.innerHTML = this._workouts.reduce(
-      (total, meal) => total + meal.calories,
-      0,
-    );
-  }
-
-  _displayCaloriesRemaining() {
-    const caloriesRemainingEl = document.getElementById('calories-remaining');
-    const progressEl = document.getElementById('calorie-progress');
-
-    const remaining = this._calorieLimit - this._totalCalories;
-    caloriesRemainingEl.innerHTML = remaining;
-
-    if (remaining <= 0) {
-      caloriesRemainingEl.parentElement.parentElement.classList.remove(
-        'bg-light',
-      );
-      caloriesRemainingEl.parentElement.parentElement.classList.add(
-        'bg-danger',
-      );
-      progressEl.classList.remove('bg-success');
-      progressEl.classList.add('bg-danger');
-    } else {
-      caloriesRemainingEl.parentElement.parentElement.classList.remove(
-        'bg-danger',
-      );
-      caloriesRemainingEl.parentElement.parentElement.classList.add('bg-light');
-      progressEl.classList.add('bg-success');
-      progressEl.classList.remove('bg-danger');
-    }
-  }
-
-  _displayCaloriesProgress() {
-    const progressEl = document.getElementById('calorie-progress');
-    const percentage = (this._totalCalories / this._calorieLimit) * 100;
-    const width = Math.min(percentage, 100);
-    progressEl.style.width = `${width}%`;
-  }
-
-  _displayNewMeal(meal) {
-    const mealsEl = document.getElementById('meal-items');
-    const mealEl = document.createElement('div');
-    mealEl.classList.add('card', 'my-2');
-    mealEl.setAttribute('data-id', meal.id);
-    mealEl.innerHTML = `
-      <div class="card-body">
-        <div class="d-flex align-items-center justify-content-between">
-          <h4 class="mx-1">${meal.name}</h4>
-          <div
-            class="fs-1 bg-primary text-white text-center rounded-2 px-2 px-sm-5"
-          >
-            ${meal.calories}
-          </div>
-          <button class="delete btn btn-danger btn-sm mx-2">
-            <i class="fa-solid fa-xmark"></i>
-          </button>
-        </div>
-      </div>
-    `;
-
-    mealsEl.appendChild(mealEl);
-  }
-
-  _displayNewWorkout(workout) {
-    const workoutsEl = document.getElementById('workout-items');
-    const workoutEl = document.createElement('div');
-    workoutEl.classList.add('card', 'my-2');
-    workoutEl.setAttribute('data-id', workout.id);
-    workoutEl.innerHTML = `
-      <div class="card-body">
-        <div class="d-flex align-items-center justify-content-between">
-          <h4 class="mx-1">${workout.name}</h4>
-          <div
-            class="fs-1 bg-secondary text-white text-center rounded-2 px-2 px-sm-5"
-          >
-            ${workout.calories}
-          </div>
-          <button class="delete btn btn-danger btn-sm mx-2">
-            <i class="fa-solid fa-xmark"></i>
-          </button>
-        </div>
-      </div>
-    `;
-
-    workoutsEl.appendChild(workoutEl);
-  }
-
-  _render() {
-    this._displayCaloriesTotal();
-    this._displayCaloriesConsumed();
-    this._displayCaloriesBurned();
-    this._displayCaloriesRemaining();
-    this._displayCaloriesProgress();
-  }
-}
-
-class Meal {
-  constructor(name, calories) {
-    this.id = Math.random().toString(16).slice(2);
-    this.name = name;
-    this.calories = calories;
-  }
-}
-
-class Workout {
-  constructor(name, calories) {
-    this.id = Math.random().toString(16).slice(2);
-    this.name = name;
-    this.calories = calories;
-  }
-}
-
-class Storage {
-  static getCalorieLimit(defaultLimit = 2000) {
-    let calorieLimit;
-    if (localStorage.getItem('calorieLimit') === null) {
-      calorieLimit = defaultLimit;
-    } else {
-      calorieLimit = +localStorage.getItem('calorieLimit');
-    }
-    return calorieLimit;
-  }
-
-  static setCalorieLimit(calorieLimit) {
-    localStorage.setItem('calorieLimit', calorieLimit);
-  }
-
-  static getTotalCalories(defaultCalories = 0) {
-    let totalCalories;
-    if (localStorage.getItem('totalCalories') === null) {
-      totalCalories = defaultCalories;
-    } else {
-      totalCalories = +localStorage.getItem('totalCalories');
-    }
-    return totalCalories;
-  }
-
-  static updateTotalCalories(calories) {
-    localStorage.setItem('totalCalories', calories);
-  }
-
-  static getMeals() {
-    let meals;
-    if (localStorage.getItem('meals') === null) {
-      meals = [];
-    } else {
-      meals = JSON.parse(localStorage.getItem('meals'));
-    }
-    return meals;
-  }
-
-  static saveMeal(meal) {
-    const meals = Storage.getMeals();
-    meals.push(meal);
-    localStorage.setItem('meals', JSON.stringify(meals));
-  }
-
-  static removeMeal(id) {
-    const meals = Storage.getMeals();
-    meals.forEach((meal, index) => {
-      if (meal.id === id) {
-        meals.splice(index, 1);
-      }
-    });
-
-    localStorage.setItem('meals', JSON.stringify(meals));
-  }
-
-  static getWorkouts() {
-    let workouts;
-    if (localStorage.getItem('workouts') === null) {
-      workouts = [];
-    } else {
-      workouts = JSON.parse(localStorage.getItem('workouts'));
-    }
-    return workouts;
-  }
-
-  static saveWorkout(workout) {
-    const workouts = Storage.getWorkouts();
-    workouts.push(workout);
-    localStorage.setItem('workouts', JSON.stringify(workouts));
-  }
-
-  static removeWorkout(id) {
-    const workouts = Storage.getMeals();
-    workouts.forEach((workout, index) => {
-      if (workout.id === id) {
-        workouts.splice(index, 1);
-      }
-    });
-
-    localStorage.setItem('workouts', JSON.stringify(workouts));
-  }
-
-  static clearAll() {
-    localStorage.removeItem('workouts');
-    localStorage.removeItem('meals');
-    localStorage.removeItem('totalCalories');
-  }
-}
 
 class App {
   constructor() {
@@ -361,7 +61,7 @@ class App {
     calories.value = '';
 
     const collapseItem = document.getElementById(`collapse-${type}`);
-    const bsCollapse = new bootstrap.Collapse(collapseItem, { toggle: true });
+    const bsCollapse = new Collapse(collapseItem, { toggle: true });
   }
 
   _removeItem(type, e) {
@@ -413,7 +113,7 @@ class App {
     limit.value = '';
 
     const modalElement = document.getElementById(`limit-modal`);
-    const modal = bootstrap.Modal.getInstance(modalElement);
+    const modal = Modal.getInstance(modalElement);
     modal.hide();
   }
 }
