+++ src/js/Meal.js
@@ -1,0 +1,9 @@
+class Meal {
+  constructor(name, calories) {
+    this.id = Math.random().toString(16).slice(2);
+    this.name = name;
+    this.calories = calories;
+  }
+}
+
+export default Meal;
