+++ src/js/Workout.js
@@ -1,0 +1,9 @@
+class Workout {
+  constructor(name, calories) {
+    this.id = Math.random().toString(16).slice(2);
+    this.name = name;
+    this.calories = calories;
+  }
+}
+
+export default Workout;
